// crm_frontend_false/src/api/api.ts
export interface Create_Order_ClientDto {
    iD_Client?: number;
    name_Client?: string | undefined;
    lastName_Client?: string | undefined;
    email_Client?: string | undefined;
    telefon?: string | undefined;
    type_technology?: string | undefined;
    model_technology?: string | undefined;
    breaking_info?: string | undefined;
}

export interface List_Order_Client {
    iD_Order?: number;
    name_Client?: string | undefined;
    lastName_Client?: string | undefined;
    email_Client?: string | undefined;
    telefon?: string | undefined;
    price?: number;
}